Update RunGame.java
SO i fixed the whole shitty issue with the weapons it wasn't fixed before. Instead of eqipping a new weapon which idk why you did that in the first place, i just changed it so that the player would be established with the weapon that they chose. Its much easier and doesnt throw and error. Trying to equip that first weapon doesnt work because it doesnt exist in inventory and yeah its shitty